Display LogEntry in a treeview for detail view.  Json converters will be added later.
